% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_survival_data.R
\name{group_surv}
\alias{group_surv}
\title{Group survival data}
\usage{
group_surv(surv_data, time_intervals, col_indexes)
}
\arguments{
\item{surv_data}{Raw data in the foarm of a data frame.}

\item{time_intervals}{A vector of time points partitioning the survival time. They define the survival time groups.}

\item{col_indexes}{A list of column indexes that specify the columns index  of the survival time,
censoring indicator, and covariates used.}
}
\value{
A data frame containing the grouped survival data.
}
\description{
This function groups censored survival data.
It assumes that the raw data includes a column of observed survival time, censoring indicator,
and some additional categorical covariates. The current version does not support contineous covariates.
}
