% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_survival_data.R
\name{interval_data}
\alias{interval_data}
\title{Extract interval-specific data}
\usage{
interval_data(lower_limit, upper_limit, ...)
}
\arguments{
\item{lower_limit}{The lower limit of the time interval}

\item{upper_limit}{The upper limit of the time interval}
}
\value{
The interval-specific data in the foarm of a data frame.
}
\description{
Extract data for observations  which experience or are censored within a specified time interval.
}
