% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_survival_data.R
\name{summarize}
\alias{summarize}
\title{Compute a summary of the grouped data}
\usage{
summarize(grouped_data, time_intervals, col_indexes = c(1, 2))
}
\arguments{
\item{grouped_data}{A data frame containing the grouped data to be summarized.}

\item{time_intervals}{A vector of time points partitioning the survival time. They define the survival time groups.}

\item{col_indexes}{A vector of column indexes for which to display the summary.}
}
\value{
The data summary.
}
\description{
Compute a summary of the grouped data
}
